Refactor Structure for better organization and maintainability
## Code Organization
- Extract Builder class to `lib/structure/builder.rb`
- Extract type coercion logic to `lib/structure/types.rb`
- Simplify main `lib/structure.rb` to focus on DSL entry point

## Type System Improvements
- Rename TypeCoercions → Types module
- Simplify API: `Types.coerce(type)` vs `TypeCoercions.coercer_for_type(type)`
- Eliminate code duplication in boolean coercion logic
- Clean method names: `boolean` vs `boolean_coercer`

## Builder Optimizations
- Reduce instance variables from 5 to 3
- Eliminate `@attributes` - deduced from `@mappings.keys`
- Eliminate `@predicate_methods` - deduced from boolean types
- Store all attributes in `@mappings` with sensible defaults

## Test Organization
- Split monolithic test file into logical groups:
  - `test_core_structure.rb` - Core DSL functionality (11 tests)
  - `test_type_coercions.rb` - Type coercion and arrays (9 tests)
  - `test_boolean_predicates.rb` - Predicate methods (5 tests)
  - `test_nested_objects.rb` - Nested parsing (4 tests)
- Simplify verbose test names for better readability
- Create `test_all.rb` to run all organized tests

## Benefits
- Better separation of concerns
- Eliminated redundant data structures
- Improved maintainability and extensibility
- Cleaner, more focused files
- Same functionality: 29 tests, 41 assertions, 100% pass rate
