+++ test/test_all.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative "test_core_structure"
+require_relative "test_type_coercions"
+require_relative "test_boolean_predicates"
+require_relative "test_nested_objects"
